requirements.txt
absl-py==0.13.0
alabaster==0.7.12
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
Babel==2.9.1
backcall==0.2.0
basepairmodels==0.2.0
biothings-client==0.2.6
brotlipy==0.7.0
cachetools==4.2.2
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.1
cloudpickle==1.6.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler==0.10.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator==5.0.9
deepdish==0.3.6
deeplift==0.6.13.0
deepTools==3.5.1
deeptoolsintervals==0.1.9
dm-tree==0.1.6
docutils==0.17.1
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
Flask==2.0.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.9.0
imagesize==1.2.0
imgkit==1.2.2
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.1.2
joblib==1.0.1
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
leidenalg==0.8.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
modisco==0.5.14.2
mseqgen==0.1.1
myvariant==1.0.0
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847908682/work
networkx==2.5.1
numexpr==2.7.3
numpy==1.19.5
numpydoc==1.1.0
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
parso==0.8.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.1.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py2bit==0.3.0
pyBigWig==0.3.18
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyfaidx==0.6.1
Pygments==2.9.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==2.4.7
pysam==0.16.0.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-igraph==0.9.6
pytz==2021.1
PyWavelets==1.1.1
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638436375034/work
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
shap==0.29.3
six==1.15.0
snowballstemmer==2.1.0
Sphinx==4.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tables==3.6.1
tenacity==8.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-probability==0.12.2
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==2.2.0
tifffile==2021.7.2
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.61.2
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing-extensions==3.7.4.3
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
